# perf: skip getting package class name inside expo module